+++ noodles-cram/src/io/writer/record/convert.rs
@@ -2,6 +2,7 @@ use std::io;
 
 use bstr::BStr;
 use noodles_core::Position;
+use noodles_fasta as fasta;
 use noodles_sam::{
     self as sam,
     alignment::{
@@ -11,10 +12,14 @@ use noodles_sam::{
 };
 
 use super::{Feature, Record};
-use crate::record::{Flags, MateFlags};
+use crate::{
+    container::compression_header::preservation_map::substitution_matrix::Base,
+    record::{Flags, MateFlags},
+};
 
 impl Record {
     pub fn try_from_alignment_record(
+        reference_sequence_repository: &fasta::Repository,
         header: &sam::Header,
         record: &dyn sam::alignment::Record,
     ) -> io::Result<Self> {
@@ -33,12 +38,32 @@ impl Record {
             record.quality_scores().iter().collect::<io::Result<_>>()?
         };
 
-        let features = cigar_to_features(
-            record.cigar().as_ref(),
-            cram_flags,
-            &sequence,
-            &quality_scores,
-        )?;
+        let reference_sequence_id = record.reference_sequence_id(header).transpose()?;
+        let alignment_start = record.alignment_start().transpose()?;
+
+        let features = if let (Some(id), Some(start)) = (reference_sequence_id, alignment_start) {
+            let (reference_sequence_name, _) = header
+                .reference_sequences()
+                .get_index(id)
+                .expect("missing reference sequence ID");
+
+            let reference_sequence = reference_sequence_repository
+                .get(reference_sequence_name)
+                .transpose()?
+                .expect("missing reference sequence");
+
+            cigar_to_features(
+                record.cigar().as_ref(),
+                reference_sequence.as_ref(),
+                cram_flags,
+                start,
+                &sequence,
+                &quality_scores,
+            )
+            .map(Some)?
+        } else {
+            None
+        };
 
         let data = record.data();
         let (data_buf, read_group_name) = get_filtered_data(data.as_ref())?;
@@ -50,9 +75,9 @@ impl Record {
         Ok(Self {
             bam_flags,
             cram_flags,
-            reference_sequence_id: record.reference_sequence_id(header).transpose()?,
+            reference_sequence_id,
             read_length: record.sequence().len(),
-            alignment_start: record.alignment_start().transpose()?,
+            alignment_start,
             read_group_id,
             name: record.name().map(|s| s.into()),
             mate_flags: MateFlags::default(),
@@ -61,7 +86,7 @@ impl Record {
             template_length: record.template_length()?,
             mate_distance: None,
             data: data_buf,
-            features,
+            features: features.unwrap_or_default(),
             mapping_quality: record.mapping_quality().transpose()?,
             sequence: sequence.into(),
             quality_scores: quality_scores.into(),
@@ -78,14 +103,18 @@ fn get_read_group_id(header: &sam::Header, read_group_name: &BStr) -> io::Result
 
 fn cigar_to_features(
     cigar: &dyn sam::alignment::record::Cigar,
+    reference_sequence: &fasta::record::Sequence,
     flags: Flags,
+    alignment_start: Position,
     sequence: &Sequence,
     quality_scores: &QualityScores,
 ) -> io::Result<Vec<Feature>> {
     use noodles_sam::alignment::record::cigar::op::Kind;
 
     let mut features = Vec::new();
-    let mut position = Position::MIN;
+
+    let mut reference_position = alignment_start;
+    let mut read_position = Position::MIN;
 
     for result in cigar.iter() {
         let op = result?;
@@ -93,113 +122,184 @@ fn cigar_to_features(
         match op.kind() {
             Kind::Match | Kind::SequenceMatch | Kind::SequenceMismatch => {
                 if op.len() == 1 {
-                    let base = sequence[position];
-                    let quality_score = quality_scores[position];
+                    let raw_reference_base = reference_sequence[reference_position];
+                    let raw_read_base = sequence[read_position];
 
-                    features.push(Feature::ReadBase {
-                        position,
-                        base,
-                        quality_score,
-                    });
+                    let quality_score = quality_scores[read_position];
+
+                    if !flags.quality_scores_are_stored_as_array() {
+                        features.push(Feature::QualityScore {
+                            position: read_position,
+                            quality_score,
+                        });
+                    }
+
+                    if !raw_reference_base.eq_ignore_ascii_case(&raw_read_base) {
+                        let feature = if let (Ok(reference_base), Ok(read_base)) = (
+                            Base::try_from(raw_reference_base),
+                            Base::try_from(raw_read_base),
+                        ) {
+                            Feature::Substitution {
+                                position: read_position,
+                                reference_base,
+                                read_base,
+                            }
+                        } else {
+                            Feature::ReadBase {
+                                position: read_position,
+                                base: raw_read_base,
+                                quality_score,
+                            }
+                        };
+
+                        features.push(feature);
+                    }
                 } else {
-                    let end = position
+                    let read_end = read_position
                         .checked_add(op.len())
                         .expect("attempt to add with overflow");
 
-                    let bases = sequence[position..end].to_vec();
-                    features.push(Feature::Bases { position, bases });
-
                     if !flags.quality_scores_are_stored_as_array() {
-                        let quality_scores = quality_scores[position..end].to_vec();
+                        let quality_scores = quality_scores[read_position..read_end].to_vec();
 
                         features.push(Feature::Scores {
-                            position,
+                            position: read_position,
                             quality_scores,
                         });
+                    }
+
+                    let reference_end = reference_position
+                        .checked_add(op.len())
+                        .expect("attempt to add with overflow");
+
+                    let reference_bases = &reference_sequence[reference_position..reference_end];
+                    let read_bases = &sequence[read_position..read_end];
+
+                    for (i, (&raw_reference_base, &raw_read_base)) in
+                        reference_bases.iter().zip(read_bases).enumerate()
+                    {
+                        if !raw_reference_base.eq_ignore_ascii_case(&raw_read_base) {
+                            let position = read_position
+                                .checked_add(i)
+                                .expect("attempt to add with overflow");
+
+                            let feature = if let (Ok(reference_base), Ok(read_base)) = (
+                                Base::try_from(raw_reference_base),
+                                Base::try_from(raw_read_base),
+                            ) {
+                                Feature::Substitution {
+                                    position,
+                                    reference_base,
+                                    read_base,
+                                }
+                            } else {
+                                let quality_score = quality_scores[read_position];
+
+                                Feature::ReadBase {
+                                    position,
+                                    base: raw_read_base,
+                                    quality_score,
+                                }
+                            };
+
+                            features.push(feature);
+                        }
                     }
                 }
             }
             Kind::Insertion => {
                 if op.len() == 1 {
-                    let base = sequence[position];
-                    features.push(Feature::InsertBase { position, base });
+                    let base = sequence[read_position];
+                    features.push(Feature::InsertBase {
+                        position: read_position,
+                        base,
+                    });
 
                     if !flags.quality_scores_are_stored_as_array() {
-                        let quality_score = quality_scores[position];
+                        let quality_score = quality_scores[read_position];
 
                         features.push(Feature::QualityScore {
-                            position,
+                            position: read_position,
                             quality_score,
                         });
                     }
                 } else {
-                    let end = position
+                    let end = read_position
                         .checked_add(op.len())
                         .expect("attempt to add with overflow");
 
-                    let bases = sequence[position..end].to_vec();
-                    features.push(Feature::Insertion { position, bases });
+                    let bases = sequence[read_position..end].to_vec();
+                    features.push(Feature::Insertion {
+                        position: read_position,
+                        bases,
+                    });
 
                     if !flags.quality_scores_are_stored_as_array() {
-                        let quality_scores = quality_scores[position..end].to_vec();
+                        let quality_scores = quality_scores[read_position..end].to_vec();
 
                         features.push(Feature::Scores {
-                            position,
+                            position: read_position,
                             quality_scores,
                         });
                     }
                 }
             }
             Kind::Deletion => features.push(Feature::Deletion {
-                position,
+                position: read_position,
                 len: op.len(),
             }),
             Kind::Skip => features.push(Feature::ReferenceSkip {
-                position,
+                position: read_position,
                 len: op.len(),
             }),
             Kind::SoftClip => {
-                let end = position
+                let end = read_position
                     .checked_add(op.len())
                     .expect("attempt to add with overflow");
 
-                let bases = &sequence[position..end];
+                let bases = &sequence[read_position..end];
 
                 features.push(Feature::SoftClip {
-                    position,
+                    position: read_position,
                     bases: bases.to_vec(),
                 });
 
                 if !flags.quality_scores_are_stored_as_array() {
                     if bases.len() == 1 {
-                        let quality_score = quality_scores[position];
+                        let quality_score = quality_scores[read_position];
 
                         features.push(Feature::QualityScore {
-                            position,
+                            position: read_position,
                             quality_score,
                         });
                     } else {
-                        let quality_scores = quality_scores[position..end].to_vec();
+                        let quality_scores = quality_scores[read_position..end].to_vec();
 
                         features.push(Feature::Scores {
-                            position,
+                            position: read_position,
                             quality_scores,
                         });
                     }
                 }
             }
             Kind::HardClip => features.push(Feature::HardClip {
-                position,
+                position: read_position,
                 len: op.len(),
             }),
             Kind::Pad => features.push(Feature::Padding {
-                position,
+                position: read_position,
                 len: op.len(),
             }),
         }
 
+        if op.kind().consumes_reference() {
+            reference_position = reference_position
+                .checked_add(op.len())
+                .expect("attempt to add with overflow");
+        }
+
         if op.kind().consumes_read() {
-            position = position
+            read_position = read_position
                 .checked_add(op.len())
                 .expect("attempt to add with overflow");
         }
@@ -246,31 +346,89 @@ mod tests {
 
     #[test]
     fn test_cigar_to_features() -> Result<(), Box<dyn std::error::Error>> {
+        let reference_sequence = fasta::record::Sequence::from(b"ACGT".to_vec());
+
         let flags = Flags::default();
+        let alignment_start = Position::MIN;
 
         let cigar: Cigar = [Op::new(Kind::Match, 1)].into_iter().collect();
         let sequence = Sequence::from(b"A");
         let quality_scores = [45].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
-        let expected = vec![Feature::ReadBase {
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![Feature::QualityScore {
             position: Position::try_from(1)?,
-            base: b'A',
             quality_score: 45,
         }];
+        assert_eq!(actual, expected);
+
+        let cigar: Cigar = [Op::new(Kind::Match, 1)].into_iter().collect();
+        let sequence = Sequence::from(b"C");
+        let quality_scores = [45].into_iter().collect();
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![
+            Feature::QualityScore {
+                position: Position::try_from(1)?,
+                quality_score: 45,
+            },
+            Feature::Substitution {
+                position: Position::try_from(1)?,
+                reference_base: Base::A,
+                read_base: Base::C,
+            },
+        ];
         assert_eq!(actual, expected);
 
         let cigar: Cigar = [Op::new(Kind::Match, 2)].into_iter().collect();
         let sequence = Sequence::from(b"AC");
         let quality_scores = [45, 35].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![Feature::Scores {
+            position: Position::try_from(1)?,
+            quality_scores: vec![45, 35],
+        }];
+        assert_eq!(actual, expected);
+
+        let cigar: Cigar = [Op::new(Kind::Match, 2)].into_iter().collect();
+        let sequence = Sequence::from(b"AT");
+        let quality_scores = [45, 35].into_iter().collect();
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
-            Feature::Bases {
-                position: Position::try_from(1)?,
-                bases: vec![b'A', b'C'],
-            },
             Feature::Scores {
                 position: Position::try_from(1)?,
                 quality_scores: vec![45, 35],
+            },
+            Feature::Substitution {
+                position: Position::try_from(2)?,
+                reference_base: Base::C,
+                read_base: Base::T,
             },
         ];
         assert_eq!(actual, expected);
@@ -280,7 +438,14 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"AC");
         let quality_scores = [45, 35].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::InsertBase {
                 position: Position::try_from(1)?,
@@ -290,10 +455,14 @@ mod tests {
                 position: Position::try_from(1)?,
                 quality_score: 45,
             },
-            Feature::ReadBase {
+            Feature::QualityScore {
                 position: Position::try_from(2)?,
-                base: b'C',
                 quality_score: 35,
+            },
+            Feature::Substitution {
+                position: Position::try_from(2)?,
+                reference_base: Base::A,
+                read_base: Base::C,
             },
         ];
         assert_eq!(actual, expected);
@@ -303,7 +472,14 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"ACG");
         let quality_scores = [45, 35, 43].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::Insertion {
                 position: Position::try_from(1)?,
@@ -313,10 +489,14 @@ mod tests {
                 position: Position::try_from(1)?,
                 quality_scores: vec![45, 35],
             },
-            Feature::ReadBase {
+            Feature::QualityScore {
                 position: Position::try_from(3)?,
-                base: b'G',
                 quality_score: 43,
+            },
+            Feature::Substitution {
+                position: Position::try_from(3)?,
+                reference_base: Base::A,
+                read_base: Base::G,
             },
         ];
         assert_eq!(actual, expected);
@@ -326,19 +506,32 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"AC");
         let quality_scores = [45, 35].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::Deletion {
                 position: Position::try_from(1)?,
                 len: 1,
             },
-            Feature::Bases {
-                position: Position::try_from(1)?,
-                bases: vec![b'A', b'C'],
-            },
             Feature::Scores {
                 position: Position::try_from(1)?,
                 quality_scores: vec![45, 35],
+            },
+            Feature::Substitution {
+                position: Position::try_from(1)?,
+                reference_base: Base::C,
+                read_base: Base::A,
+            },
+            Feature::Substitution {
+                position: Position::try_from(2)?,
+                reference_base: Base::G,
+                read_base: Base::C,
             },
         ];
         assert_eq!(actual, expected);
@@ -348,16 +541,27 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"A");
         let quality_scores = [45].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::ReferenceSkip {
                 position: Position::try_from(1)?,
                 len: 1,
             },
-            Feature::ReadBase {
+            Feature::QualityScore {
                 position: Position::try_from(1)?,
-                base: b'A',
                 quality_score: 45,
+            },
+            Feature::Substitution {
+                position: Position::try_from(1)?,
+                reference_base: Base::C,
+                read_base: Base::A,
             },
         ];
         assert_eq!(actual, expected);
@@ -367,7 +571,14 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"AC");
         let quality_scores = [45, 35].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::SoftClip {
                 position: Position::try_from(1)?,
@@ -377,10 +588,14 @@ mod tests {
                 position: Position::try_from(1)?,
                 quality_score: 45,
             },
-            Feature::ReadBase {
+            Feature::QualityScore {
                 position: Position::try_from(2)?,
-                base: b'C',
                 quality_score: 35,
+            },
+            Feature::Substitution {
+                position: Position::try_from(2)?,
+                reference_base: Base::A,
+                read_base: Base::C,
             },
         ];
         assert_eq!(actual, expected);
@@ -390,7 +605,14 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"ACG");
         let quality_scores = [45, 35, 43].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::SoftClip {
                 position: Position::try_from(1)?,
@@ -400,10 +622,14 @@ mod tests {
                 position: Position::try_from(1)?,
                 quality_scores: vec![45, 35],
             },
-            Feature::ReadBase {
+            Feature::QualityScore {
                 position: Position::try_from(3)?,
-                base: b'G',
                 quality_score: 43,
+            },
+            Feature::Substitution {
+                position: Position::try_from(3)?,
+                reference_base: Base::A,
+                read_base: Base::G,
             },
         ];
         assert_eq!(actual, expected);
@@ -413,15 +639,21 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"A");
         let quality_scores = [45].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::HardClip {
                 position: Position::try_from(1)?,
                 len: 1,
             },
-            Feature::ReadBase {
+            Feature::QualityScore {
                 position: Position::try_from(1)?,
-                base: b'A',
                 quality_score: 45,
             },
         ];
@@ -432,15 +664,21 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"A");
         let quality_scores = [45].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::Padding {
                 position: Position::try_from(1)?,
                 len: 1,
             },
-            Feature::ReadBase {
+            Feature::QualityScore {
                 position: Position::try_from(1)?,
-                base: b'A',
                 quality_score: 45,
             },
         ];
@@ -452,26 +690,72 @@ mod tests {
     #[test]
     fn test_cigar_to_features_with_quality_scores_stored_as_array()
     -> Result<(), Box<dyn std::error::Error>> {
+        let reference_sequence = fasta::record::Sequence::from(b"ACGT".to_vec());
+
         let flags = Flags::QUALITY_SCORES_ARE_STORED_AS_ARRAY;
+        let alignment_start = Position::MIN;
 
         let cigar: Cigar = [Op::new(Kind::Match, 1)].into_iter().collect();
         let sequence = Sequence::from(b"A");
         let quality_scores = [45].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
-        let expected = vec![Feature::ReadBase {
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![];
+        assert_eq!(actual, expected);
+
+        let cigar: Cigar = [Op::new(Kind::Match, 1)].into_iter().collect();
+        let sequence = Sequence::from(b"C");
+        let quality_scores = [45].into_iter().collect();
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![Feature::Substitution {
             position: Position::try_from(1)?,
-            base: b'A',
-            quality_score: 45,
+            reference_base: Base::A,
+            read_base: Base::C,
         }];
         assert_eq!(actual, expected);
 
         let cigar: Cigar = [Op::new(Kind::Match, 2)].into_iter().collect();
         let sequence = Sequence::from(b"AC");
         let quality_scores = [45, 35].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
-        let expected = vec![Feature::Bases {
-            position: Position::try_from(1)?,
-            bases: vec![b'A', b'C'],
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![];
+        assert_eq!(actual, expected);
+
+        let cigar: Cigar = [Op::new(Kind::Match, 2)].into_iter().collect();
+        let sequence = Sequence::from(b"AT");
+        let quality_scores = [45, 35].into_iter().collect();
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![Feature::Substitution {
+            position: Position::try_from(2)?,
+            reference_base: Base::C,
+            read_base: Base::T,
         }];
         assert_eq!(actual, expected);
 
@@ -480,16 +764,23 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"AC");
         let quality_scores = [45, 35].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::InsertBase {
                 position: Position::try_from(1)?,
                 base: b'A',
             },
-            Feature::ReadBase {
+            Feature::Substitution {
                 position: Position::try_from(2)?,
-                base: b'C',
-                quality_score: 35,
+                reference_base: Base::A,
+                read_base: Base::C,
             },
         ];
         assert_eq!(actual, expected);
@@ -499,16 +790,23 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"ACG");
         let quality_scores = [45, 35, 43].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::Insertion {
                 position: Position::try_from(1)?,
                 bases: vec![b'A', b'C'],
             },
-            Feature::ReadBase {
+            Feature::Substitution {
                 position: Position::try_from(3)?,
-                base: b'G',
-                quality_score: 43,
+                reference_base: Base::A,
+                read_base: Base::G,
             },
         ];
         assert_eq!(actual, expected);
@@ -518,15 +816,28 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"AC");
         let quality_scores = [45, 35].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::Deletion {
                 position: Position::try_from(1)?,
                 len: 1,
             },
-            Feature::Bases {
+            Feature::Substitution {
                 position: Position::try_from(1)?,
-                bases: vec![b'A', b'C'],
+                reference_base: Base::C,
+                read_base: Base::A,
+            },
+            Feature::Substitution {
+                position: Position::try_from(2)?,
+                reference_base: Base::G,
+                read_base: Base::C,
             },
         ];
         assert_eq!(actual, expected);
@@ -536,16 +847,23 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"A");
         let quality_scores = [45].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::ReferenceSkip {
                 position: Position::try_from(1)?,
                 len: 1,
             },
-            Feature::ReadBase {
+            Feature::Substitution {
                 position: Position::try_from(1)?,
-                base: b'A',
-                quality_score: 45,
+                reference_base: Base::C,
+                read_base: Base::A,
             },
         ];
         assert_eq!(actual, expected);
@@ -555,16 +873,23 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"AC");
         let quality_scores = [45, 35].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::SoftClip {
                 position: Position::try_from(1)?,
                 bases: vec![b'A'],
             },
-            Feature::ReadBase {
+            Feature::Substitution {
                 position: Position::try_from(2)?,
-                base: b'C',
-                quality_score: 35,
+                reference_base: Base::A,
+                read_base: Base::C,
             },
         ];
         assert_eq!(actual, expected);
@@ -574,16 +899,23 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"ACG");
         let quality_scores = [45, 35, 43].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
         let expected = vec![
             Feature::SoftClip {
                 position: Position::try_from(1)?,
                 bases: vec![b'A', b'C'],
             },
-            Feature::ReadBase {
+            Feature::Substitution {
                 position: Position::try_from(3)?,
-                base: b'G',
-                quality_score: 43,
+                reference_base: Base::A,
+                read_base: Base::G,
             },
         ];
         assert_eq!(actual, expected);
@@ -593,18 +925,18 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"A");
         let quality_scores = [45].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
-        let expected = vec![
-            Feature::HardClip {
-                position: Position::try_from(1)?,
-                len: 1,
-            },
-            Feature::ReadBase {
-                position: Position::try_from(1)?,
-                base: b'A',
-                quality_score: 45,
-            },
-        ];
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![Feature::HardClip {
+            position: Position::try_from(1)?,
+            len: 1,
+        }];
         assert_eq!(actual, expected);
 
         let cigar: Cigar = [Op::new(Kind::Pad, 1), Op::new(Kind::Match, 1)]
@@ -612,18 +944,18 @@ mod tests {
             .collect();
         let sequence = Sequence::from(b"A");
         let quality_scores = [45].into_iter().collect();
-        let actual = cigar_to_features(&cigar, flags, &sequence, &quality_scores)?;
-        let expected = vec![
-            Feature::Padding {
-                position: Position::try_from(1)?,
-                len: 1,
-            },
-            Feature::ReadBase {
-                position: Position::try_from(1)?,
-                base: b'A',
-                quality_score: 45,
-            },
-        ];
+        let actual = cigar_to_features(
+            &cigar,
+            &reference_sequence,
+            flags,
+            alignment_start,
+            &sequence,
+            &quality_scores,
+        )?;
+        let expected = vec![Feature::Padding {
+            position: Position::try_from(1)?,
+            len: 1,
+        }];
         assert_eq!(actual, expected);
 
         Ok(())